# src/main/kotlin/tech/pacifici/patronus/controller/PatronusPostControllerApi.kt
package tech.pacifici.patronus.controller

import org.springframework.http.HttpStatus
import org.springframework.http.ResponseEntity
import org.springframework.web.bind.annotation.RestController
import tech.pacifici.patronus.api.PatronusPostsApi
import tech.pacifici.patronus.model.CommentRequest
import tech.pacifici.patronus.model.CommentResponse
import tech.pacifici.patronus.model.PostRequest
import tech.pacifici.patronus.model.PostResponse
import java.util.UUID

/**
 * REST controller implementation for the Patronus API.
 * This class provides endpoints for managing users, posts, comments, and friendships.
 * It implements the [PatronusApi] interface and handles HTTP requests related to the core features of the Patronus application.
 */
@RestController
class PatronusPostControllerApi : PatronusPostsApi {
    /**
     * Creates a new comment.
     * @param id The Id of the comment
     * @param xRequestId The x-request-id of the call
     * @param xCorrelationId The correlation Id associated to the flow o calls
     * @param commentRequest The request body containing the details of the comment to create.
     * @return A [ResponseEntity] containing the created [CommentResponse] and HTTP status.
     */
    override fun createComment(
        id: UUID,
        xRequestId: UUID,
        xCorrelationId: UUID,
        commentRequest: CommentRequest,
    ): ResponseEntity<CommentResponse> {
        // TODO: implement the logic here
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    /**
     * Creates a new post.
     *
     * @param xRequestId The x-request-id of the call
     * @param xCorrelationId The correlation Id associated to the flow o calls
     * @param postRequest The request body containing the details of the post to create.
     * @return A [ResponseEntity] containing the created [PostResponse] and HTTP status.
     */
    override fun createPost(
        xRequestId: UUID,
        xCorrelationId: UUID,
        postRequest: PostRequest,
    ): ResponseEntity<PostResponse> {
        // TODO: Implement your business logic here
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    /**
     * Deletes a post by its ID.
     *
     * @param id The ID of the post to delete.
     * @param xRequestId The x-request-id of the call
     * @param xCorrelationId The correlation Id associated to the flow o calls
     * @return A [ResponseEntity] with HTTP status indicating the result of the operation.
     */
    override fun deletePostById(
        id: UUID,
        xRequestId: UUID,
        xCorrelationId: UUID,
    ): ResponseEntity<Unit> {
        // TODO: Implement your business logic here
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    /**
     * Retrieves a post by its ID.
     *
     * @param id The ID of the post to retrieve.
     * @param xRequestId The x-request-id of the call
     * @param xCorrelationId The correlation Id associated to the flow o calls
     * @return A [ResponseEntity] containing the [PostResponse] and HTTP status.
     */
    override fun getPostById(
        id: UUID,
        xRequestId: UUID,
        xCorrelationId: UUID,
    ): ResponseEntity<PostResponse> {
        // TODO: Implement your business logic here
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    /**
     * Retrieves all posts.
     *
     * @param xRequestId The x-request-id of the call
     * @param xCorrelationId The correlation Id associated to the flow o calls
     * @return A [ResponseEntity] containing a list of [PostResponse] and HTTP status.
     */
    override fun getPosts(
        xRequestId: UUID,
        xCorrelationId: UUID,
    ): ResponseEntity<List<PostResponse>> {
        // TODO: Implement your business logic here
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    /**
     * Updates a post by its ID.
     *
     * @param id The ID of the post to update.
     * @param xRequestId The x-request-id of the call
     * @param xCorrelationId The correlation Id associated to the flow o calls
     * @param postRequest The request body containing the updated details of the post.
     * @return A [ResponseEntity] containing the updated [PostResponse] and HTTP status.
     */
    override fun updatePostById(
        id: UUID,
        xRequestId: UUID,
        xCorrelationId: UUID,
        postRequest: PostRequest,
    ): ResponseEntity<PostResponse> {
        // TODO: Implement your business logic here
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }
}